Popover now responds to ^{N,P,J,K}
Typically typing control and n in a TextView advances to the next line, and control-p moves to the previous line. This commit introduces similar functionality to the script selection Popover. It supports both Emacs-like key chords, C-n and C-p, as well as Vi-like alternatives, C-j and C-k. Similar functionality is offered by Dash and Launchbar.